by sam

+++ Online-Plant-Nursery/src/main/java/com/cg/online_plant_nursery/dao/CartDAO.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.cg.online_plant_nursery.entity.Cart;
 
 @Repository
-public interface CartDAO extends JpaRepository<Cart, Integer>{
+public interface CartDAO extends JpaRepository<Cart, Long>{
 	
 	@Query(value = "from Cart cart where cart.id=?1")
 	 public Cart getCartById(long id);
