+++ Online-Plant-Nursery/src/main/java/com/cg/online_plant_nursery/dao/GardenDecorDAO.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.cg.online_plant_nursery.entity.GardenDecor;
 
 @Repository
-public interface GardenDecorDAO extends JpaRepository<GardenDecor,Integer> {
+public interface GardenDecorDAO extends JpaRepository<GardenDecor,Long> {
 
 	@Query(value = "from GardenDecor decor where decor.id=?1")
 	public GardenDecor getDecorById(long id);
